+++ app/mailers/booked_mailer.rb
@@ -15,9 +15,19 @@ class BookedMailer < ActionMailer::Base
 
   def booking_status(booking)
     @booking = booking
-    @area = booking.post.area
+    @area = booking.post.area_available
     @poster = booking.poster.full_name
     subject = "Cancellation booking"
+    mail(
+      :subject => subject,
+      :to => booking.poster.email
+      )
+  end
+
+  def booking_payment(booking)
+    @booking = booking
+    @poster = booking.poster.full_name
+    subject = "Transfer of funds"
     mail(
       :subject => subject,
       :to => booking.poster.email